requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Mail==0.9.1
python-dotenv==1.0.0
itsdangerous==2.1.2
Werkzeug==2.3.6
requests==2.31.0
